Address CodeRabbit feedback on PR #727
- Move DiscussionThreadNotification + NotificationEnvelope into
  supabase/functions/_shared/FunctionTypes.d.ts and import them in
  notification-queue-processor/index.ts.
- Suppress students-channel broadcast for instructors_only threads in
  20260518120000_discussion_threads_broadcast_root_transition.sql.
- Fail fast on grader rename setup step in discussion-threads.test.tsx.
